% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_variables.R
\name{add_clc_bfr_sums_capillary}
\alias{add_clc_bfr_sums_capillary}
\title{Title}
\usage{
add_clc_bfr_sums_capillary(.data, .bound = "medium", .add_PCBTOT = TRUE)
}
\arguments{
\item{.data}{}

\item{.bound}{}

\item{.add_PCBSUM}{}
}
\description{
Title
}
