% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report_helper_functions.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Title}
\usage{
add_missing_columns(.data, .template, .select_template_columns = TRUE)
}
\arguments{
\item{.select_template_columns}{}
}
\description{
Title
}
