% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_api_data.R
\name{add_nrm_code}
\alias{add_nrm_code}
\title{Title}
\usage{
add_nrm_code(
  .data,
  codes_path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse")
)
}
\arguments{
\item{.data}{}

\item{codes_path}{}
}
\description{
Title
}
