% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_helpers.R
\name{add_smnh_variable_names}
\alias{add_smnh_variable_names}
\title{Title}
\usage{
add_smnh_variable_names(.data, .cols = everything())
}
\arguments{
\item{.data}{}

\item{.cols}{}
}
\description{
Title
}
