% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report_helper_functions.R
\name{convert_cols_to_character}
\alias{convert_cols_to_character}
\title{Title}
\usage{
convert_cols_to_character(.data, round_numeric = TRUE, digits = 5)
}
\arguments{
\item{.data}{}

\item{round_numeric}{}

\item{digits}{}
}
\description{
Title
}
