% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report_helper_functions.R
\name{export_to_template}
\alias{export_to_template}
\title{Title}
\usage{
export_to_template(
  .data,
  template_path = system.file("extdata", "miljogifter_leveransmall.xlsx", package =
    "MoCiS2.wtse"),
  export_path = "miljogifter_leveransmall_ifylld.xlsx",
  overwrite = FALSE,
  start_col = 2,
  start_row = 6,
  col_names = FALSE
)
}
\arguments{
\item{col_names}{}
}
\description{
Title
}
