% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_code_lists.R
\name{get_code_list}
\alias{get_code_list}
\title{Title}
\usage{
get_code_list(
  list_name = c("kembiofys", "provmetadata", "data_matvarde", "bestallare", "provdata",
    "inspire", "prc", "current"),
  exclude_sheets = c("version"),
  clean_names = FALSE,
  current = "codelist.xlsx"
)
}
\arguments{
\item{list_name}{}

\item{exclude_sheets}{}

\item{clean_names}{}

\item{current}{}
}
\description{
Title
}
