% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_code_lists.R
\name{get_code_list_rmdk}
\alias{get_code_list_rmdk}
\title{Title}
\usage{
get_code_list_rmdk(
  list_name,
  clean_names = FALSE,
  list_file = "rmdk_kodlistor_lankar.xlsx"
)
}
\arguments{
\item{list_file}{}
}
\description{
Title
}
