% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_code_lists.R
\name{get_codes_prc}
\alias{get_codes_prc}
\title{Title}
\usage{
get_codes_prc(
  path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse"),
  sheet = "PARAMETRAR",
  columns = c("NRM_PARAMETERKOD", "PROV_BERED", "PROVKARL", "ANALYS_MET", "ANALYS_INSTR")
)
}
\arguments{
\item{sheet}{}
}
\description{
Title
}
