% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_code_lists.R
\name{get_codes_species}
\alias{get_codes_species}
\title{Title}
\usage{
get_codes_species(
  .data = NULL,
  path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse"),
  sheet = "ARTER",
  fuzzyjoin = TRUE,
  max_dist = 2
)
}
\arguments{
\item{max_dist}{}
}
\description{
Title
}
