% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_code_lists.R
\name{get_codes_substances}
\alias{get_codes_substances}
\title{Title}
\usage{
get_codes_substances(
  path = system.file("extdata", "codelist.xlsx", package = "MoCiS2"),
  sheet = "PARAMETRAR",
  columns = c("NRM_PARAMETERKOD", "PARAMETERNAMN", "UNIK_PARAMETERKOD", "ENHET",
    "MATOSAKERHET_ENHET", "PROV_LAGR")
)
}
\arguments{
\item{path}{}

\item{sheet}{}
}
\description{
Title
}
