% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_code_lists.R
\name{get_codes_substances_methods}
\alias{get_codes_substances_methods}
\title{Title}
\usage{
get_codes_substances_methods(
  path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse"),
  sheet = "PARAMETRAR",
  columns = c("NRM_PARAMETERKOD", "LABB", "PROV_BERED", "PROVKARL", "ANALYS_MET",
    "UTFOR_LABB", "ANALYS_INSTR")
)
}
\arguments{
\item{path}{}

\item{sheet}{}

\item{columns}{}
}
\description{
Title
}
