% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_mocis_stations}
\alias{get_mocis_stations}
\title{Title}
\usage{
get_mocis_stations(
  path = system.file("extdata", "codelist.xlsx", package = "MoCiS2")
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
