% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_variables.R
\name{import_clc_bfr_new_template}
\alias{import_clc_bfr_new_template}
\title{Title}
\usage{
import_clc_bfr_new_template(
  path,
  summary_fn = NULL,
  ackr_sum = NULL,
  format = c("long", "wide"),
  ...
)
}
\arguments{
\item{format}{}
}
\description{
Title
}
