% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{moc_join_lab}
\alias{moc_join_lab}
\title{Title}
\usage{
moc_join_lab(
  .x,
  negative_for_nondetect = TRUE,
  codes_path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse"),
  has_provid = TRUE,
  na_values = c("-99.99", "N/A", "-99.9"),
  remove_na = TRUE,
  fuzzyjoin = TRUE,
  max_dist = 2,
  remove_na_measurement = TRUE,
  remove_na_parameter_name = TRUE
)
}
\arguments{
\item{max_dist}{}
}
\description{
Title
}
