% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report.R
\name{moc_join_SGU}
\alias{moc_join_SGU}
\title{Title}
\usage{
moc_join_SGU(
  biodata,
  analysdata,
  add_bio_pars = TRUE,
  codes_path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse")
)
}
\arguments{
\item{add_bio_pars}{}
}
\description{
Title
}
