% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{moc_read_lab}
\alias{moc_read_lab}
\title{Extracts information from lab protocol}
\usage{
moc_read_lab(
  path,
  negative_for_nondetect = TRUE,
  codes_path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse"),
  has_provid = TRUE,
  na_values = c("-99.99", "N/A", "-99.9"),
  remove_na = TRUE,
  fuzzyjoin = 2,
  max_dist = 2,
  join = TRUE,
  dates_as_character = TRUE,
  summary_fn = NULL,
  ...
)
}
\arguments{
\item{has_provid}{This should be FALSE for old sheets without a column for NRM provid}
}
\description{
Extracts information from lab protocol
}
