% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{moc_read_prc}
\alias{moc_read_prc}
\title{Title}
\usage{
moc_read_prc(
  prc_path,
  codes_path = system.file("extdata", "codelist_prc.xls", package = "MoCiS2"),
  enc = "latin1"
)
}
\arguments{
\item{codes_path}{}
}
\description{
Title
}
