% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_join_data_shared_functions.R
\name{moc_read_warnings_base_message}
\alias{moc_read_warnings_base_message}
\title{Title}
\usage{
moc_read_warnings_base_message(data, .filter_col, .pull_col)
}
\arguments{
\item{.pull_col}{}
}
\description{
Title
}
