% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report.R
\name{moc_write_SGU}
\alias{moc_write_SGU}
\title{Title}
\usage{
moc_write_SGU(
  data,
  sheet,
  file = NULL,
  template_path = system.file("extdata", "miljogifter-leveransmall_2024.xlsx", package =
    "MoCiS2.wtse"),
  program = "none",
  convert_to_character = TRUE
)
}
\arguments{
\item{program}{}
}
\description{
Title
}
