% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocis_extract_official_stats.R
\name{mocis_extract_official_stats}
\alias{mocis_extract_official_stats}
\title{Title}
\usage{
mocis_extract_official_stats(contaminants, sites, species, run_date, dir)
}
\arguments{
\item{dir}{}
}
\description{
Title
}
