% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_lab_file2}
\alias{read_lab_file2}
\title{Title}
\usage{
read_lab_file2(
  path,
  variable,
  sheet,
  .has_provid = TRUE,
  na_values = c("-99.99", "N/A"),
  pivot_longer = TRUE
)
}
\arguments{
\item{sheet}{}
}
\description{
Title
}
