% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_lab_file_date}
\alias{read_lab_file_date}
\title{Title}
\usage{
read_lab_file_date(
  path,
  sheet = "date of analysis",
  .has_provid = TRUE,
  na_values = c("-99.99", "N/A"),
  dates_as_character = TRUE,
  pivot_longer = TRUE
)
}
\arguments{
\item{sheet}{}
}
\description{
Title
}
