% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_api_data.R
\name{sgu_data_to_wide}
\alias{sgu_data_to_wide}
\title{Title}
\usage{
sgu_data_to_wide(
  .data,
  .vars,
  codes_path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse")
)
}
\arguments{
\item{.data}{}

\item{.vars}{}

\item{codes_path}{}
}
\description{
Title
}
