% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{unpool}
\alias{unpool}
\title{Given PROV_KOD_ORIGINAL, splits into a table of individual PROV_KOD_ORIGINAL and corresponding pooled PROV_KOD_ORIGINAL}
\usage{
unpool(PROV_KOD_ORIGINAL)
}
\arguments{
\item{PROV_KOD_ORIGINAL}{An NRM ACCNR}
}
\value{
A tibble
}
\description{
Given PROV_KOD_ORIGINAL, splits into a table of individual PROV_KOD_ORIGINAL and corresponding pooled PROV_KOD_ORIGINAL
}
\examples{
unpool("C2016/00937-00939")
}
