% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report_wtse.R
\name{wtse_format_SGU}
\alias{wtse_format_SGU}
\title{Title}
\usage{
wtse_format_SGU(
  data,
  sheet,
  template_path = system.file("extdata", "miljogifter-leveransmall_2024.xlsx", package =
    "MoCiS2.wtse"),
  convert_to_character = TRUE,
  round_numeric = TRUE,
  digits = 5
)
}
\arguments{
\item{program}{}
}
\description{
Title
}
