% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report_wtse.R
\name{wtse_join_SGU}
\alias{wtse_join_SGU}
\title{Title}
\usage{
wtse_join_SGU(
  biodata,
  analysdata,
  has_pools = FALSE,
  add_bio_pars = TRUE,
  codes_path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2"),
  measurement_cols = c("TOTV", "TOTL", "BRED", "FOSL", "FOSV", "SKTJ", "SKLV", "SKTJ",
    "SKVO", "SKLI", "UTTO"),
  remove_na_biodata = FALSE
)
}
\arguments{
\item{add_bio_pars}{}
}
\description{
Title
}
