% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report_helper_functions.R
\name{export_to_template}
\alias{export_to_template}
\title{Title}
\usage{
export_to_template(
  .data,
  template_path = system.file("extdata", "miljogifter-leveransmall_2024.xlsx", package =
    "MoCiS2"),
  export_path = "miljogifter-leveransmall-ifylld.xlsx",
  overwrite = FALSE,
  start_col = 2,
  start_row = 6,
  col_names = FALSE
)
}
\arguments{
\item{.data}{}

\item{template_path}{}

\item{export_path}{}

\item{overwrite}{}

\item{start_col}{}

\item{start_row}{}

\item{col_names}{}
}
\description{
Title
}
