% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGU_report_helper_functions.R
\name{read_xlsx_sgu_template}
\alias{read_xlsx_sgu_template}
\title{Title}
\usage{
read_xlsx_sgu_template(
  path,
  sheet,
  ul_data = c(6, 2),
  ul_names = c(1, 2),
  lr_names = c(1, NA),
  offset = 1,
  col_types = sgu_template_col_types(sheet),
  ...
)
}
\arguments{
\item{lr_names}{}
}
\description{
Title
}
